refactor(ci-tests): the claim detector splits from the rules it serves
The file reached 1014 lines and the craft gate refused the push, which was
the right call: it had grown to answer two questions at once.

`uniquenessclaimsdetector_test.go` now holds which comments are claims,
which declaration each belongs to, and the key that tells two of them
apart. `uniquenessclaims_test.go` keeps the rules and their fixtures. A
reader asking "what does this gate refuse" wants the second; a reader asking
"would it see MY comment" wants the first, and neither should scroll past
the other to get there.

Both are skipped by the sweep, because both spell out the phrases they hunt
for — the skip is a SET now rather than one path, and removing the second
entry fails.
